% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RapidRoughSets-package.R
\docType{package}
\name{RapidRoughSets}
\alias{RapidRoughSets}
\alias{RapidRoughSets-package}
\title{Efficient data analysis with methods deriving from the rough sets theory}
\description{
Efficient data analysis with methods deriving from the rough sets theory
}

